Use editor build config for list of scenes
It's tedious to copy and syncronize the list of scenes between both
Unity's build config and the different build configs I can make in
Anchor. Morever, I don't really have a need for this feature, so I'm
going to remove it for now, at least until such need arises again.