Improved build packaging
Instead of naming the folder the application is in for standalone builds
with the same name of the folder used in the Builds folder, use the
product name instead.

Similarly, for web builds, it now packages the zip properly in a format
that itch expects (it must contain a root-level index.html file).